Add files via upload
Sorting and analyzing fleet inventory data Using Pivot Table.

Objectives
•	Clean The Data.
•	Format data as a table.
•	Create a Pivot Table and use fields to arrange data in a Pivot Table.
•	Perform calculations using Pivot Table data.

DATA CLEANING.
1.Column widths: 
Sorting out the widths of all columns so that the data is clearly visible in all cells.
 2. Empty rows: 
Using the Filter feature to look for blanks and removing all empty rows from the data. 
3. Duplicate records: 
Using either the Conditional Formatting or Remove Duplicates feature to look for and remove any duplicated records from the data. 
4. Spelling: 
The original source file data has not been checked for errors in the spelling. I Checked for spelling mistakes in the data and fixed them. 
5. Whitespace: 
I Used the Find and Replace feature to remove all double-spaces from the data. 
6. Department names: 
The department names are not correctly entered and they are now split over two columns in the data. I Used Flash Fill to reduce the department names to just one column, and then removed any unnecessary columns.
CREATING A PIVOT TABLE:
1. Format the data as a table: Used the Format as Table option to format the data as a table. 
2. Use AutoSum to calculate values: 
I Used AutoSum to find the following values for column ‘C’ and recorded each of the values: 
•	SUM
•	AVERAGE 
•	MIN 
•	MAX 
•	COUNT 
3. Create a Pivot Table: 
Used the PivotTable feature to create a pivot table that displays the Department field in the Rows section, and the Equipment Count in the Values section, so that the pivot table displays the sum of equipment count by department. 
4. Sort the pivot table data: 
Used the Sort by Value setting on the pivot table to sort it in descending order by the sum of equipment count. 
5. Make two more pivot tables exactly the same as task 3: Follow the same steps performed in Tasks 3 and 4 to create two more identical pivot tables so that you end up with 3 worksheets that contain identical pivot tables. 
6. Analyze data in the pivot table: 
Use the PivotTable Fields pane to manipulate and analyze data in the two copied pivot table as follows: In pivot table 2 add the Equipment Class field below the Department field so that the different vehicle types appear under each department with their respective counts. Collapse all fields except the top one - Transportation 
In pivot table 3 add the Equipment Class field above the Department field so that the different vehicle types appear first, with the different departments listed underneath each vehicle type with their respective counts. Collapse all fields except the top one - CUV